Memory improvement
New: Pointer.h, holds a pointer to objects, allowing for objects to be destroyed even though there are still pointers towards it, freeing up a lot of paused space at the cost of a few bytes per Object.

new: PointerControlBlock.h, works together with Pointer.h to act as an inbetween step, all Pointer classes will point to a control block, destroying an object will clear out the object memory, leaving only the control block behind for the pointers

change: removed as  many unnecessary inclusions as possible

change: moved the pending distructable handling to Application instead of Scene, this removes the need for a scene reference to each object.